gregisyourslackid/requirements.txt
slack-sdk>=3.20.0
slack-bolt
requests>=2.28.0
python-dotenv>=0.21.0
